feat(lexBox): Add streak tracking and refactor UI sheets
Implement streak statistics logic and reorganize presentation components.

- **Stats Implementation**:
  - Added `currentStreak` and `maximumStreak` tracking to `LexBoxStatsRepo` and the `LexBoxStatistics` interface.
  - Implemented logic to increment streaks upon winning, reset streaks if a day is missed or lost, and revert streaks if a winning word is erased.
- **Refactoring**:
  - Extracted `_LexBoxStatsSheet` and `_LexBoxTutorialSheet` from `game_layout.dart` into separate files: `stats_sheet.dart` and `tutorial_sheet.dart`.
  - Updated `LexBoxStatsSheet` to visualize current and maximum streak data alongside existing stats.
- **Fixes**:
  - Updated the win condition check (`_isWin`) to be case-insensitive.
