+++ lib/data/lexBox/implementation/stats.dart
@@ -9,6 +9,8 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
   static const _lettersOfTheDayLengthField = 'lettersOfTheDayLength';
   static const _userDataField = 'userData';
   static const _winCountField = 'winCount';
+  static const _currentStreakField = 'currentStreak';
+  static const _maximumStreakField = 'maximumStreak';
   static const _lastPlayedMatchDateField = 'lastPlayedDate';
   static const _wordsSubmittedTodayField = 'lastGuessedWords';
   static final DateFormat _dateFormat = DateFormat('dd/MM/yyyy');
@@ -24,6 +26,8 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
     var userData =
         userDataValuesDoc.exists ? userDataValuesDoc.value as Map : {};
     var winCount = _getIntegerStatistic(userData, _winCountField);
+    var currentStreak = _getIntegerStatistic(userData, _currentStreakField);
+    var maximumStreak = _getIntegerStatistic(userData, _maximumStreakField);
     List<String> lastSubmittedWords = [];
     DateTime? lastPlayedMatchDay;
     if (userData.containsKey(_lastPlayedMatchDateField)) {
@@ -40,6 +44,8 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
     }
     return LexBoxStatsRepo._(
         winCount: winCount,
+        currentStreak: currentStreak,
+        maximumStreak: maximumStreak,
         lettersOfTheDay: lettersOfTheDay,
         wordsSubmittedToday: lastSubmittedWords,
         userId: userId,
@@ -49,21 +55,30 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
 
   @override
   Future reCalculate() async {
-    if (_lastCompletedMatchDay != null) {
-      final daysDifference =
-          initializedDateTime.numberOfDaysInBetween(_lastCompletedMatchDay!);
-      if (daysDifference > 0) {
-        if (_isWin(wordsSubmittedToday)) {
-          winCount++;
-        }
+    if (_lastCompletedMatchDay != null &&
+        _lastCompletedMatchDay!.numberOfDaysInBetween(initializedDateTime) >
+            0) {
+      // Check if user had won on the last played day
+      final hadWonLastDay = _isWin(wordsSubmittedToday);
+      // Should reset streak if more than 1 day passed, or if 1 day passed without winning
+      var shouldResetStreak =
+          _lastCompletedMatchDay!.numberOfDaysInBetween(initializedDateTime) >
+                  1 ||
+              !hadWonLastDay;
+
+      final updateData = <String, Object?>{
+        _wordsSubmittedTodayField: null,
+        _lastPlayedMatchDateField: null,
+        if (shouldResetStreak) _currentStreakField: 0,
+      };
+
+      await _userDataReference.update(updateData).then((_) {
         wordsSubmittedToday.clear();
         _lastCompletedMatchDay = null;
-        await _userDataReference.update({
-          _winCountField: winCount,
-          _lastPlayedMatchDateField: null,
-          _wordsSubmittedTodayField: null,
-        });
-      }
+        if (shouldResetStreak) {
+          currentStreak = 0;
+        }
+      });
     }
   }
 
@@ -84,16 +99,47 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
   final DateTime initializedDateTime;
 
   @override
+  int currentStreak;
+
+  @override
+  int maximumStreak;
+
+  @override
   Future<bool> trySubmitWord(String word) async {
     var didSubmitWord = false;
-    await _userDataReference.update({
-      _wordsSubmittedTodayField: wordsSubmittedToday.toList()
-        ..add(word.toLowerCase()),
+    final updatedWords = wordsSubmittedToday.toList()..add(word.toLowerCase());
+
+    // Check if this word completes the game (uses all letters)
+    final willWin = _isWin(updatedWords);
+
+    final updateData = <String, dynamic>{
+      _wordsSubmittedTodayField: updatedWords,
       _lastPlayedMatchDateField: _dateFormat.format(initializedDateTime),
-    }).then((_) {
+    };
+
+    // If this word causes a win, update streak and win count
+    if (willWin) {
+      final newCurrentStreak = currentStreak + 1;
+      final newMaximumStreak =
+          newCurrentStreak > maximumStreak ? newCurrentStreak : maximumStreak;
+      updateData[_winCountField] = winCount + 1;
+      updateData[_currentStreakField] = newCurrentStreak;
+      updateData[_maximumStreakField] = newMaximumStreak;
+    }
+
+    await _userDataReference.update(updateData).then((_) {
       didSubmitWord = true;
       wordsSubmittedToday.add(word.toLowerCase());
       _lastCompletedMatchDay ??= initializedDateTime;
+
+      // Update local streak values if won
+      if (willWin) {
+        winCount++;
+        currentStreak++;
+        if (currentStreak > maximumStreak) {
+          maximumStreak = currentStreak;
+        }
+      }
     }).onError((_, __) {
       didSubmitWord = false;
     });
@@ -105,15 +151,38 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
   Future<bool> eraseLastWord() async {
     if (wordsSubmittedToday.isEmpty) return false;
     var didEraseWord = false;
+
+    // Check if currently won before erasing
+    final isCurrentlyWon = _isWin(wordsSubmittedToday);
     final updatedWords = wordsSubmittedToday.toList()..removeLast();
-    await _userDataReference.update({
+    final willStillBeWon = _isWin(updatedWords);
+
+    final updateData = <String, dynamic>{
       _wordsSubmittedTodayField: updatedWords.isEmpty ? null : updatedWords,
-      if (updatedWords.isEmpty) _lastPlayedMatchDateField: null,
-    }).then((_) {
+    };
+
+    if (updatedWords.isEmpty) {
+      updateData[_lastPlayedMatchDateField] = null;
+    }
+
+    // If erasing this word removes the win, revert the win and streak
+    if (isCurrentlyWon && !willStillBeWon) {
+      final newCurrentStreak = currentStreak > 0 ? currentStreak - 1 : 0;
+      updateData[_winCountField] = winCount > 0 ? winCount - 1 : 0;
+      updateData[_currentStreakField] = newCurrentStreak;
+    }
+
+    await _userDataReference.update(updateData).then((_) {
       didEraseWord = true;
       wordsSubmittedToday.removeLast();
       if (wordsSubmittedToday.isEmpty) {
         _lastCompletedMatchDay = null;
+      }
+
+      // Update local state if win was reverted
+      if (isCurrentlyWon && !willStillBeWon) {
+        if (winCount > 0) winCount--;
+        if (currentStreak > 0) currentStreak--;
       }
     }).onError((_, __) {
       didEraseWord = false;
@@ -135,8 +204,8 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
       .child(_userId);
 
   bool _isWin(Iterable<String> words) {
-    var allLetters = lettersOfTheDay.split('').toSet();
-    var usedLetters = words.expand((w) => w.split('')).toSet();
+    var allLetters = lettersOfTheDay.toLowerCase().split('').toSet();
+    var usedLetters = words.expand((w) => w.toLowerCase().split('')).toSet();
     return allLetters.every(usedLetters.contains);
   }
 
@@ -162,6 +231,8 @@ class LexBoxStatsRepo extends LexBoxStatsModifier {
 
   LexBoxStatsRepo._(
       {required this.winCount,
+      required this.currentStreak,
+      required this.maximumStreak,
       required this.lettersOfTheDay,
       required this.wordsSubmittedToday,
       required String userId,
